build: Edit build structure, add resolveJsonModule to tsconfig
Impotus for this change was to fix @typescript/no-var-requires by adding
resolveJsonModule to tsconfig. This adds package.json to the root of
build dir, so source files are placed by the compiler in build/src.

Then realized it is cleaner for build to contain all files actually
needed. This way we can include just build in the npm package.